+++ src/components/Molecules/EventCard/EventCard.tsx
@@ -1,11 +1,14 @@
 import React from 'react';
 
-import { CardWrapper, CardWrapperProps } from '../../Atoms';
+import { CardWrapper, CardWrapperProps, Tag } from '../../Atoms';
+import classNames from 'classnames';
 
 export interface EventCardProps extends Omit<CardWrapperProps, 'innerClassNames'> {
   date: string | number | Date;
   title: string;
   months?: [string, ...string[]] & { length: 12 };
+  isComplete?: boolean;
+  completeLabel?: string;
 }
 
 const defaultMonths: [string, ...string[]] & { length: 12 } = [
@@ -28,13 +31,19 @@ const EventCard = ({
   date,
   color,
   children,
+  completeLabel,
   months = defaultMonths,
+  isComplete = false,
   ...props
 }: EventCardProps) => {
   const dateObject = new Date(date);
 
+  const wrapperClasses = classNames('idsk-event-card', {
+    'idsk-event-card--complete': isComplete
+  });
+
   return (
-    <CardWrapper color={color} innerClassNames="idsk-event-card" {...props}>
+    <CardWrapper color={isComplete ? '#BDBDBD' : color} innerClassNames={wrapperClasses} {...props}>
       <div className="idsk-event-card__date">
         <time dateTime={dateObject.toISOString()} className="idsk-event-card__day">
           {dateObject.getDate()}
@@ -42,7 +51,12 @@ const EventCard = ({
         <p className="idsk-event-card__month">{months[dateObject.getMonth()]}</p>
       </div>
       <div>
-        <div className="idsk-event-card__title">{title}</div>
+        <div className="idsk-event-card__title">
+          {title}
+          {isComplete && completeLabel && (
+            <Tag className="idsk-event-card__complete-tag" size="small" label={completeLabel} />
+          )}
+        </div>
         {children}
       </div>
     </CardWrapper>