+++ src/svgImages/Logos/index.ts
@@ -1,1 +1,2 @@
 export { default as LogoPrivate } from './LogoPrivate';
+export { default as LogoImage } from './LogoImage';